requirements.txt
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.0
pillow>=9.5.0
numpy>=1.24.0
scipy>=1.10.0
librosa>=0.10.0
soundfile>=0.12.0
matplotlib>=3.7.0
plotly>=5.15.0
# Optional: xformers for memory optimization (install manually if needed)
# xformers>=0.0.22 